+++ OUDS/Core/Themes/Sosh/Sources/SoshTheme.swift
@@ -1,0 +1,21 @@
+//
+// Software Name: OUDS iOS
+// SPDX-FileCopyrightText: Copyright (c) Orange SA
+// SPDX-License-Identifier: MIT
+// 
+// This software is distributed under the MIT license,
+// the text of which is available at https://opensource.org/license/MIT/
+// or see the "LICENSE" file for more details.
+// 
+// Authors: See CONTRIBUTORS.txt
+// Software description: A SwiftUI components library with code examples for Orange Unified Design System 
+//
+
+import Foundation
+import OUDSThemesCommons
+
+/// Overrides some colors using values defined in extension of `ColorRawTokens` in this current module.
+open class SoshTheme: OUDSTheme {
+
+    // For clarity reasons, please override OUDSTheme properties in extensions
+}